rename: infinite 경로 수정
